third_party/WebKit/LayoutTests/media/avtrack/audio-track-properties.html

<!doctype html>
<html>
    <head>
        <title>AudioTrack properties read from input stream metadata</title>
        <script src="../../resources/testharness.js"></script>
        <script src="../../resources/testharnessreport.js"></script>
        <script src="../media-file.js"></script>
    </head>
    <body>
        <div id="log"></div>
        <script>
            async_test(function(t)
            {
                var e = document.createElement("audio");
                e.src = "../content/test.webm";

                e.onloadedmetadata = t.step_func_done(function()
                {
                    assert_equals(e.audioTracks.length, 1, "audioTracks.length");
                    assert_true(e.audioTracks[0].enabled, "audioTrack.enabled");
                    assert_equals(e.audioTracks[0].id, "2", "audioTrack.id");
                    assert_equals(e.audioTracks[0].kind, "main", "audioTrack.kind");
                    assert_equals(e.audioTracks[0].label, "", "audioTrack.label");
                    assert_equals(e.audioTracks[0].language, "", "audioTrack.language");
                });
            });
        </script>
    </body>
</html>
